backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Document processing
PyMuPDF>=1.23.0
openpyxl>=3.1.2
python-pptx>=0.6.23
python-docx>=1.1.0
pandas>=2.1.0

# AI / NLP
openai>=1.10.0
anthropic>=0.18.0
tiktoken>=0.5.0

# Text processing
rapidfuzz>=3.6.0
regex>=2023.12.0

# Database (for later phases)
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.0
PyYAML>=6.0.1